Movie

+++ app/src/main/java/com/qoolqas/moviedb/model/similiar/SimiliarResultsItem.kt
@@ -35,7 +35,7 @@ data class SimiliarResultsItem(
 	val popularity: Double? = null,
 
 	@field:SerializedName("vote_average")
-	val voteAverage: Int? = null,
+	val voteAverage: Double? = null,
 
 	@field:SerializedName("id")
 	val id: Int? = null,
